+++ src/aks-agent/azext_aks_agent/_params.py
@@ -2,6 +2,8 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
+from azure.cli.core.commands.parameters import get_enum_type
+
 
 # pylint: disable=too-many-statements,too-many-lines
 def load_arguments(self, _):
@@ -23,6 +25,12 @@ def load_arguments(self, _):
             help="Name of the managed cluster.",
         )
         c.argument(
+            "namespace",
+            options_list=["--namespace"],
+            help="The Kubernetes namespace where the AKS Agent is deployed. Required for cluster mode.",
+            required=False,
+        )
+        c.argument(
             "max_steps",
             type=int,
             default=40,
@@ -39,6 +47,12 @@ def load_arguments(self, _):
             "no_interactive",
             help="Disable interactive mode. When set, the agent will not prompt for input and will run in batch mode.",
             action="store_true",
+        )
+        c.argument(
+            "mode",
+            arg_type=get_enum_type(["cluster", "client"]),
+            help="The mode decides how the agent is deployed.",
+            default="cluster",
         )
         c.argument(
             "no_echo_request",
@@ -73,6 +87,7 @@ def load_arguments(self, _):
             options_list=["--name", "-n"],
             help="Name of the managed cluster.",
         )
+
     with self.argument_context("aks agent-cleanup") as c:
         c.argument(
             "resource_group_name",
@@ -83,4 +98,16 @@ def load_arguments(self, _):
             "cluster_name",
             options_list=["--name", "-n"],
             help="Name of the managed cluster.",
+        )
+        c.argument(
+            "namespace",
+            options_list=["--namespace"],
+            help="The Kubernetes namespace where the AKS Agent is deployed. Required for cluster mode.",
+            required=False,
+        )
+        c.argument(
+            "mode",
+            arg_type=get_enum_type(["cluster", "client"]),
+            help="The mode decides how the agent is deployed.",
+            default="cluster",
         )